webpage display is good
I modified Rhea's query function so we can use it for the back end. I put the modified py file inside of the query function file.

I just made sure that the one I pushed is going to work if you guy pull it. Now, the only problem is that the csv file is not going to be read in when calling the modified query function in the main app because the path I specified when read in the final dataset  is my local path. I think there must be a specific file that we can put the csv file in so when calling the query function inside of main app, it can be detected with no problem. I could not figure this out, so I had to specify where the csv is located. If you guys change the path to your own local path, the webpage will display properly.